+++ include/adc_manager.h
@@ -4,4 +4,5 @@ namespace AdcManager
 {
     void begin();
     float readDutVoltage();
+    float readAlarmSenseVoltage();
 }