+++ src/test_defs.cpp
@@ -6,8 +6,10 @@ const char *testIdToString(TestId testId)
     {
     case TestId::POWER_TEST:
         return "Power Test";
-    case TestId::ALARM_TEST:
-        return "Alarm Test";
+    case TestId::ALARM_POSITIVE_TEST:
+        return "Alarm Positive Test";
+    case TestId::ALARM_NEGATIVE_TEST:
+        return "Alarm Negative Test";
     case TestId::OPEN_CIRCUIT_TEST:
         return "Open Circuit Test";
     case TestId::SHORT_CIRCUIT_TEST: